factor out code common across receivers
